This layout probably seems best
Also moved some code to documentation.rs to avoid making help.rs massive
