+++ crates/nu-cli/src/commands.rs
@@ -30,6 +30,7 @@ pub(crate) mod date;
 pub(crate) mod debug;
 pub(crate) mod default;
 pub(crate) mod do_;
+mod documentation;
 pub(crate) mod drop;
 pub(crate) mod du;
 pub(crate) mod each;
